pyproject.toml
[project]
name = "hello-world-genai"
version = "0.2.0" # for APAN5560 course Assignment 2
description = "FastAPI application with Bigram Text Generation, Word Embeddings, and CIFAR-10 Image Classification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "jupyter>=1.1.1",
    "pydantic>=2.12.5",
    "spacy>=3.7.0",
    "scikit-learn>=1.4.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
]
